YAMLTEST: Refactoring - separating yaml test harness from the actual tests.
