+++ fhirpath/src/test/java/au/csiro/pathling/fhirpath/yaml/YamlFhirpathTest.java
@@ -1,5 +1,7 @@
 package au.csiro.pathling.fhirpath.yaml;
 
+import au.csiro.pathling.test.yaml.YamlSpec;
+import au.csiro.pathling.test.yaml.YamlSpecTestBase;
 import jakarta.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Disabled;
