+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/YamlSupport.java
@@ -1,4 +1,4 @@
-package au.csiro.pathling.fhirpath.yaml;
+package au.csiro.pathling.test.yaml;
 
 import static java.util.Objects.isNull;
 import static java.util.Objects.requireNonNull;
